project/requirements.txt
gym
autopep8
sockjs-tornado
prometheus_client
matplotlib
IPython
tqdm
pyglet
torch-tb-profiler
tensorboard
torchsummary
cached_property

yfinance
yahoo-fin
psycopg2-binary
sqlalchemy
pytorch_lightning